Fix buf generator's target path
