+++ buf.gen.yaml
@@ -3,16 +3,16 @@ managed:
   enabled: true
 plugins:
   - plugin: buf.build/community/neoeinstein-prost:v0.2.2
-    out: tycho/src/pb
+    out: tycho-indexer/src/pb
     opt: file_descriptor_set=false
 
   - plugin: buf.build/community/neoeinstein-tonic:v0.2.2
-    out: tycho/src/pb
+    out: tycho-indexer/src/pb
     opt:
       - no_server=true
 
   - plugin: buf.build/community/neoeinstein-prost-crate:v0.3.1
-    out: tycho/src/pb
+    out: tycho-indexer/src/pb
     opt:
       - include_file=pb.rs
       - no_features
