+++ tycho-indexer/src/pb/pb.rs
@@ -1,6 +1,9 @@
 // @generated
 pub mod sf {
+    // @@protoc_insertion_point(attribute:sf.substreams)
     pub mod substreams {
+        include!("sf.substreams.rs");
+        // @@protoc_insertion_point(sf.substreams)
         pub mod internal {
             // @@protoc_insertion_point(attribute:sf.substreams.internal.v2)
             pub mod v2 {
@@ -13,6 +16,15 @@ pub mod sf {
             pub mod v2 {
                 include!("sf.substreams.rpc.v2.rs");
                 // @@protoc_insertion_point(sf.substreams.rpc.v2)
+            }
+        }
+        pub mod sink {
+            pub mod service {
+                // @@protoc_insertion_point(attribute:sf.substreams.sink.service.v1)
+                pub mod v1 {
+                    include!("sf.substreams.sink.service.v1.rs");
+                    // @@protoc_insertion_point(sf.substreams.sink.service.v1)
+                }
             }
         }
         // @@protoc_insertion_point(attribute:sf.substreams.v1)
